Editing  migrations

+++ database/migrations/2022_09_23_054449_create_subscriptions_table.php
@@ -14,15 +14,15 @@ return new class extends Migration
     public function up()
     {
         Schema::create('subscriptions', function (Blueprint $table) {
-            $table->string("id")->primary();
-            $table->string("type");
-            $table->boolean("is_free");
-            $table->date("start_date");
-            $table->date("end_date");
-            $table->unsignedInteger("plan_id");
+            $table->string('id')->primary();
+            $table->string('type');
+            $table->boolean('is_free');
+            $table->date('start_date');
+            $table->date('end_date');
+            $table->unsignedBigInteger('plan_id');
             $table->timestamps();
 
-            $table->foreign("plan_id")->references("id")->on("plans")->onDelete("cascade");
+            $table->foreign('plan_id')->references('id')->on('plans')->onDelete('cascade');
         });
     }
