+++ database/migrations/2022_09_25_214354_create_car_features_table.php
@@ -15,17 +15,17 @@ return new class extends Migration
     {
         Schema::create('car_features', function (Blueprint $table) {
             $table->id();
-            $table->string("country_of_manufacture");
-            $table->string("fuel");
-            $table->string("color");
-            $table->enum("car_type",["new","used","both"]);
-            $table->integer("manufacturing_year");
-            $table->float("maximum_speed");
-            $table->float("mileage");
-            $table->unsignedInteger("car_id");
+            $table->string('country_of_manufacture');
+            $table->string('fuel');
+            $table->string('color');
+            $table->enum('car_type',['new','used','both']);
+            $table->integer('manufacturing_year');
+            $table->float('maximum_speed');
+            $table->float('mileage');
+            $table->unsignedBigInteger('car_id');
             $table->timestamps();
             
-            $table->foreign("car_id")->references("id")->on("cars")->onDelete("cascade");
+            $table->foreign('car_id')->references('id')->on('cars')->onDelete('cascade');
         });
     }
